Rename reindex reingest; remove diff code
